requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
openai==1.11.1
scholarly==1.7.11
langchain==0.1.0
tqdm==4.66.1
fake-useragent==1.4.0
python-dotenv==1.0.0
arxiv==2.0.0